requirements.txt
# Function dependencies, for example:
# package>=version
pillow
google-cloud-storage
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.10.0+cpu
https://download.pytorch.org/whl/cpu/torchvision-0.11.0%2Bcpu-cp37-cp37m-linux_x86_64.whl